% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{loadQconfig}
\alias{loadQconfig}
\title{Clase R6 para la gestion de las preguntas de moodle}
\usage{
loadQconfig(file)
}
\description{
creación de bancos de preguntas para moodle con clases R

Manual: \url{https://mcstllns.github.io/moodleQ/}
}
