% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questions.R
\name{question}
\alias{question}
\title{Clase R6 para la gestion de las preguntas de moodle}
\description{
creación de bancos de preguntas para moodle con clases R

Manual: \url{https://mcstllns.github.io/moodleQ/}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{type}}{El tipo de pregunta}

\item{\code{prop}}{Contiene todas las propiedades de las preguntas}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{question$new()}}
\item \href{#method-set}{\code{question$set()}}
\item \href{#method-remove}{\code{question$remove()}}
\item \href{#method-print}{\code{question$print()}}
\item \href{#method-errorHandler}{\code{question$errorHandler()}}
\item \href{#method-xml_parser}{\code{question$xml_parser()}}
\item \href{#method-xml}{\code{question$xml()}}
\item \href{#method-removeanswer}{\code{question$removeanswer()}}
\item \href{#method-addanswer}{\code{question$addanswer()}}
\item \href{#method-addfile}{\code{question$addfile()}}
\item \href{#method-addDataset}{\code{question$addDataset()}}
\item \href{#method-clone}{\code{question$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{question$new(type = NA, confuser = 0, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set"></a>}}
\if{latex}{\out{\hypertarget{method-set}{}}}
\subsection{Method \code{set()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{question$set(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-remove"></a>}}
\if{latex}{\out{\hypertarget{method-remove}{}}}
\subsection{Method \code{remove()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{question$remove(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{question$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-errorHandler"></a>}}
\if{latex}{\out{\hypertarget{method-errorHandler}{}}}
\subsection{Method \code{errorHandler()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{question$errorHandler(id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-xml_parser"></a>}}
\if{latex}{\out{\hypertarget{method-xml_parser}{}}}
\subsection{Method \code{xml_parser()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{question$xml_parser(l, nombre)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-xml"></a>}}
\if{latex}{\out{\hypertarget{method-xml}{}}}
\subsection{Method \code{xml()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{question$xml(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-removeanswer"></a>}}
\if{latex}{\out{\hypertarget{method-removeanswer}{}}}
\subsection{Method \code{removeanswer()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{question$removeanswer(num)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-addanswer"></a>}}
\if{latex}{\out{\hypertarget{method-addanswer}{}}}
\subsection{Method \code{addanswer()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{question$addanswer(
  text,
  fraction = 0,
  format = self$prop$format$.value,
  tolerance = 0.001,
  tolerancetype = 2,
  correctanswerformat = 1,
  correctanswerlength = 3
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-addfile"></a>}}
\if{latex}{\out{\hypertarget{method-addfile}{}}}
\subsection{Method \code{addfile()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{question$addfile(drivePath, name = NULL, quizPath = "/")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-addDataset"></a>}}
\if{latex}{\out{\hypertarget{method-addDataset}{}}}
\subsection{Method \code{addDataset()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{question$addDataset(df, name = NA, status = "shared", type = "calculated")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{question$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
