% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiz.R
\name{quiz}
\alias{quiz}
\title{Clase R6 para la gestion de los cuestionarios}
\description{
Gestiona los cuestionarios
}
\details{
Crea cuestionarios compuestos por preguntas moodleQ
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{quiz$new()}}
\item \href{#method-add}{\code{quiz$add()}}
\item \href{#method-print}{\code{quiz$print()}}
\item \href{#method-xml}{\code{quiz$xml()}}
\item \href{#method-remove}{\code{quiz$remove()}}
\item \href{#method-save_xml}{\code{quiz$save_xml()}}
\item \href{#method-clone}{\code{quiz$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{quiz$new(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add"></a>}}
\if{latex}{\out{\hypertarget{method-add}{}}}
\subsection{Method \code{add()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{quiz$add(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{quiz$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-xml"></a>}}
\if{latex}{\out{\hypertarget{method-xml}{}}}
\subsection{Method \code{xml()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{quiz$xml(file)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-remove"></a>}}
\if{latex}{\out{\hypertarget{method-remove}{}}}
\subsection{Method \code{remove()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{quiz$remove(qids)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-save_xml"></a>}}
\if{latex}{\out{\hypertarget{method-save_xml}{}}}
\subsection{Method \code{save_xml()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{quiz$save_xml(file)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{quiz$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
